Feat: Migrate volume to its own component
